Work done to login/register/dashboard/logout
Testing and making sure that login was testing not just username, but password as well, which is working now. Logout now clears the session and any cookies that are associated with it. Dashboard is only accessible if a session is active and the session params correspond to data in the database. Register is still functioning as usual.